+++ CVServiceMapAPI/Public/Get-CVServiceMapMachinesSummary.ps1
@@ -19,6 +19,9 @@
 
   Limitations and Known Issues
   ----------------------------
+  - ShowAllVmsStatus param is a bit flawed : 
+    - Only shows in current Subscription where OMS is 
+    - Only shows active machines so if machine switched off , may show as Not installed
   
   Backlog 
   --------
@@ -28,7 +31,7 @@
   v1.00 Andy Ball 17/02/2017 Base Version
   v1.01 Andy Ball 19/02/2017 Add timestamp field /param 
   v1.02 Andy Ball 21/02/2017 Add MachineId to output
-  
+  v1.03 Andy Ball 22/02/2017 Add ShowsAllVMsStatus param 
  
  .Parameter OMSWorkspaceName
   Name of OMS Workspace 
@@ -38,6 +41,9 @@
 
  .Parameter SubscriptionName
  Subscription where OMS is located. Looks in current Subsription if null
+
+ .ShowAllVMsStatus
+ False by default, if true does a Get-AzureRMVM in $SubscriptionName to attempt to show Servers where no Service Map data. See limitations above
 
  .Parameter LocalTimeStamp
  When specified gets info for this date time. Pass local time , will be converted to UTC time into API
@@ -55,13 +61,17 @@ Function Get-CVServiceMapMachinesSummary
             [Parameter(Mandatory = $true, Position = 0)]  [string] $OMSWorkspaceName  	,
             [Parameter(Mandatory = $true, Position = 1)]  [string] $ResourceGroupName ,
             [Parameter(Mandatory = $false, Position = 2)]  [string] $SubscriptionName , 
-            [Parameter(Mandatory = $false, Position = 3)]  [string] $LocalTimeStamp 
+            [Parameter(Mandatory = $false, Position = 3)]  [boolean] $ShowAllVMsStatus = $false , 
+            [Parameter(Mandatory = $false, Position = 4)]  [string] $LocalTimeStamp 
 
         )
 
     # ie silly date in the past so we get everything , 
     # ToDo not 100% sure how date works , maybe should just do offset from Now().UTC>AddHours(-2) or something ..
     # $uriSuffix = "/machines/?api-version=2015-11-01-preview&live=false&timestamp=2017-02-17T09:57:56.9366303Z"
+
+
+    $Resultset = @()
 
     $TimeStampSuffix = ""
     If([string]::IsNullOrWhiteSpace($LocalTimeStamp) -eq $false)
@@ -81,9 +91,8 @@ Function Get-CVServiceMapMachinesSummary
         {
             $ret = Get-CVServiceMapWrapper -OMSWorkspaceName $OMSWorkspaceName -ResourceGroupName $ResourceGroupName -SubscriptionName $SubscriptionName -URISuffix $uriSuffix
         }
-    $ret.value | Select @{Name = "ComputerName" ; Expression = {$_.Properties.ComputerName}}, 
-                        @{Name = "MachineId" ; Expression = {$_.id}},
-                        @{Name = "timestamp" ; Expression = {$_.Properties.timestamp}}, 
+    $Resultset = $ret.value | Select @{Name = "ComputerName" ; Expression = {$_.Properties.ComputerName}}, 
+                          @{Name = "timestamp" ; Expression = {$_.Properties.timestamp}}, 
                         @{Name = "MachineName" ; Expression = {$_.name}}, 
                         @{Name = "FirstIPAddress" ; Expression = {$_.Properties.networking.ipv4Interfaces[0].ipAddress}}, 
                         @{Name = "UTCBootTime" ; Expression = {$_.Properties.bootTime}} , 
@@ -94,7 +103,35 @@ Function Get-CVServiceMapMachinesSummary
                         @{Name  = "VMType" ;  Expression = {$_.Properties.virtualMachine.virtualMachineType}},
                         @{Name  = "MemoryMB" ;  Expression = {$_.Properties.resources.physicalMemory}},
                         @{Name  = "CPUs" ;  Expression = {$_.Properties.resources.cpus}},
-                        @{Name = "OS" ; Expression = {$_.Properties.operatingSystem.fullName}} | Sort ComputerName
+                        @{Name = "OS" ; Expression = {$_.Properties.operatingSystem.fullName}} ,
+                        @{Name = "MachineId" ; Expression = {$_.id}}
 
+
+    # Create a row for each VM in Subscription that doesnt appear in ServiceMap results 
+    If ($ShowAllVMsStatus)
+        {
+            Write-Host "Running Get-AzureRMVM so can check for inactive / non-installed machines."
+            $AllVMs = Get-AzureRMVM 
+            $VMsWithoutServiceMap = $AllVMs | Where {$_.Name -notin $Resultset.ComputerName}
+            ForEach ($MissingVM in $VMsWithoutServiceMap)
+                {
+                    $ResultSet += $Host | Select @{Name = "ComputerName" ; Expression = {$MissingVM.Name}}, 
+                          @{Name = "timestamp" ; Expression = {"N\A"}}, 
+                        @{Name = "MachineName" ; Expression = {"N\A"}}, 
+                        @{Name = "FirstIPAddress" ; Expression = {"N\A"}}, 
+                        @{Name = "UTCBootTime" ; Expression = {"N\A"}} , 
+                        @{Name = "TimeZoneDiff" ; Expression = {"N\A"}},
+                        @{Name = "AgentVersion" ; Expression = {"Not installed"}},
+                        @{Name = "AgentRevision" ; Expression = {"Not Installed"}}, 
+                        @{Name = "AgentRebootStatus" ; Expression = {"N\A"}},
+                        @{Name  = "VMType" ;  Expression = {"N\A"}},
+                        @{Name  = "MemoryMB" ;  Expression = {"N\A"}},
+                        @{Name  = "CPUs" ;  Expression = {"N\A"}},
+                        @{Name = "OS" ; Expression = {"N\A"}} ,
+                        @{Name = "MachineId" ; Expression = {"N\A"}}
+                }
+        }
+    # Finally return
+    $Resultset | Sort ComputerName
 }